refactor: rename server actions for member management; change useActionStatus file location.
 useActionStatus was accidentally created inside the server actions folder.
